Make a neuronlist object from two separate files

Make a neuronlist object from two separate files

read_nl_from_parts(datapath, dfpath = NULL, package = NULL, ...)

Arguments

datapath

Path to the data object

dfpath

Path to the data.frame object (constructed from datapath when NULL, see details)

package

Character vector naming a package whose extdata directory will be sought (with find_extdata) and prepended to the two input paths.

...

Additional arguments passd to find_extdata

Value

a neuronlist object

Details

It is expected that you will use this in an R source file within the data folder of a package. See Examples for more information.

If dfpath is missing, it will be inferred from datapath according to the following pattern:

myblob.rda main data file

myblob.df.rda metdata file

See also

Examples

if (FALSE) {
# you could use the following in a file
# data/make_data.R
delayedAssign('pns', read_nl_from_parts('pns.rds', package='testlazyneuronlist'))
# based on objects created by
save_nl_in_parts(pns)
# which would make:
# - inst/extdata/pns.rds
# - inst/extdata/pns.df.rds
}